refactor: 패키지 이동
- http 연결과 관련된 것은 coyote로
- http 요청 처리 관련된 것은 catalina로
