+++ tomcat/src/main/java/org/apache/catalina/view/ViewResolver.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.view;
+package org.apache.catalina.view;
 
 import java.io.File;
 import java.io.IOException;
