+++ tomcat/src/main/java/org/apache/coyote/connector/Connector.java
@@ -1,4 +1,4 @@
-package org.apache.catalina.connector;
+package org.apache.coyote.connector;
 
 import org.apache.coyote.http11.Http11Processor;
 import org.slf4j.Logger;
